Improve usage with uppercase (append) registers
The uppercase letter macro registers append what is written to them to
the lowercase variant when you save.

The initial content of the uppercase register is the same as the
lowercase one.

But if you want to use this macro editor for appending reasons, it makes
no sense to paste the contents in the uppercase buffers.

Fixed by only pasting if argument is a lowercase char.